Melhorando a exibição

+++ install-postgresql.sh
@@ -2,62 +2,91 @@
 
 echo "### Atualizando o sistema ###"
 yum update -y
-
-echo "### Verificando versões do postgresql ###"
-
-yum module list postgresql
-
+echo "---------------------------"
+echo "                        Pronto!!!                       "
 echo ""
-
+echo "### Verificando versões do postgresql ###"
+yum module list postgresql
+echo "---------------------------"
+echo "                        Pronto!!!                       "
+echo ""
+echo ""
 echo -n "Por favor, selecione a versão a ser instalada: "
 read version
+echo "---------------------------"
 echo ""
-
+echo "                        Pronto!!!                       "
+echo ""
 echo "### Configurando a versão selecionada como padrão ###"
 yum module enable postgresql:$version -y
+echo "---------------------------"
 echo ""
-
+echo "                        Pronto!!!                       "
+echo ""
+echo ""
 echo "### Instalando ###"
 yum install postgresql-server -y
+echo "---------------------------"
 echo ""
-
+echo "                        Pronto!!!                       "
+echo ""
 echo "### Iniciando e habilitando o serviço no boot ###"
 #Verificando se o diretório /var/lib/pgsql/data/ está vazio
-if [ -d /var/lib/pgsql/data/* ]
+if [ -z /var/lib/pgsql/data/ ]
 then
-	# Se o diretório não estiver vazio
-	echo "O postgresql já está funcionando!!!"
-else
 	# Se o diretório estiver vazio
 	service postgresql initdb
 	systemctl start postgresql
 	systemctl enable postgresql
-fi
-echo ""
 
+else
+	# Se o diretório não estiver vazio
+	echo "O postgresql já está funcionando!!!"
+fi
+echo "---------------------------"
+echo ""
+echo "                        Pronto!!!                       "
+echo ""
 echo "### Configurando o postgresql para acesso remoto ###"
 echo "------"
+echo ""
 echo "### Alterando o arquivo pg_hba.conf ###"
 sed -i 's/127.0.0.1\/32/0.0.0.0\/0/g' /var/lib/pgsql/data/pg_hba.conf
 sed -i 's/ident/trust/g' /var/lib/pgsql/data/pg_hba.conf
 echo "------"
+echo ""
 echo "### Alterando o arquivo /var/lib/pgsql/data/postgresql.conf ###"
 sed -i 's/#listen_addresses/listen_addresses/g' /var/lib/pgsql/data/postgresql.conf
 sed -i 's/localhost/\*/g' /var/lib/pgsql/data/postgresql.conf
 sed -i 's/#port = 5432/port = 5432/g' /var/lib/pgsql/data/postgresql.conf
+echo "---------------------------"
 echo ""
-
+echo "                        Pronto!!!                       "
+echo ""
 echo "### Alterando a senha do usuário postgre ###"
 sudo -u postgres psql -c "ALTER USER postgres WITH PASSWORD 'postgres'"
+echo "---------------------------"
+echo ""
+echo "                        Pronto!!!                       "
 echo ""
 echo "### Reiniciando o serviço do postgresql ###"
 systemctl restart postgresql
+echo "---------------------------"
 echo ""
-
+echo "                        Pronto!!!                       "
+echo ""
 echo "### Configurando firewall ###"
-systemctl stop firewalld
-systemctl disable firewalld
-systemctl mask firewalld
+echo""
+# Verificando se o forewalld está ativo
+firewalld_active=$(systemctl status firewalld | grep inactive | wc -l)
+if [ $firewalld_active -eq 1 ]
+then
+	echo "O firewalld já está desativado"
+else
+	systemctl stop firewalld
+	systemctl disable firewalld
+	systemctl mask firewalld
+fi
 
 systemctl start nftables
 systemctl enable nftables
@@ -99,3 +128,10 @@ sed -i 's/\/etc\/nftables\/main.nft/\/etc\/nftables\/firewall.nft/g' /etc/syscon
 sed -i 's/#include/include/g' /etc/sysconfig/nftables.conf
 
 systemctl restart nftables
+echo "---------------------------"
+echo ""
+echo "                        Pronto!!!                       "
+echo ""
+echo "Nftables configurado"
+echo ""
+echo "### Pronto ###"
